move announce day to extension module
