+++ duckbot/cogs/announce_day/__init__.py
@@ -1,0 +1,5 @@
+from .announce_day import AnnounceDay
+
+
+def setup(bot):
+    bot.add_cog(AnnounceDay(bot))
