+++ duckbot/cogs/announce_day/phrases.py
@@ -1,0 +1,91 @@
+days = {
+    0: {
+        "names": [
+            "the day of the moon",
+            "Moonday",
+            "Monday",
+        ],
+        "templates": [
+            "Screw this, it is {0}.",
+        ],
+    },
+    1: {
+        "names": [
+            "Tyr's day",
+            "Tuesday",
+            "Dndndndndndndnd",
+        ],
+        "templates": [
+            "Yo, one of my favourite days ever: {0}.",
+        ],
+    },
+    2: {
+        "names": [
+            "Odin's day",
+            "Wednesday",
+            "Wednesday, my dudes",
+            "hump day",
+            "Ness' wedding day",
+            "https://www.youtube.com/watch?v=du-TY1GUFGk",
+        ],
+        "templates": [],
+    },
+    3: {
+        "names": [
+            "Thor's day",
+            "Thorsday",
+            "civ day",
+            "Thursday",
+        ],
+        "templates": [],
+    },
+    4: {
+        "names": [
+            "Frigga's day",
+            "Friday, Friday, gotta get down on Friday",
+            "Friday",
+        ],
+        "templates": [],
+    },
+    5: {
+        "names": [
+            "Tom's Movie Night",
+            "Saturn's day",
+            "Saturday",
+        ],
+        "templates": [
+            "WEEKEND ALERT: {0}!",
+        ],
+    },
+    6: {
+        "names": [
+            "the day of the sun",
+            "Sunday",
+            "Sunday, Sunday, Sunday",
+        ],
+        "templates": [
+            "WEEKEND ALERT: {0}!",
+        ],
+    },
+}
+
+templates = [
+    "Today is {0}.",
+    "Today: {0}. Tomorrow: {1}.",
+    "Brothers, the day is {0}.",
+    "Yoooooo, today is {0}! Brother.",
+    "The day is {0}. Prepare yourself.",
+    "What if I said to you that in fact, today is not {0}? I'd be lying.",
+    "What if I said to you that in fact, today is not {1}? I'd be right. It's {0}.",
+    "Rejoice, for today is {0}.",
+    "Don't get it twisted, it's {0}.",
+    "{0}",
+    "It {0}.",
+    "Today is {0}, I hope you have your pants on.",
+    "Gentlemen, it is with great pleasure that I inform you. Today is {0}.",
+    "Huh, it's {0}, but it feels more like {1}.",
+    "The day of the week according to the Gregorian calendar is {0}.",
+    "Beep boop: {0}.",
+    "I take no pleasure in announcing that today is {0}, for I am a robot.",
+    "I take no displeasure in announcing that today is {0}, for I am a robot.",
+]
